(service):feat 更新 Milvus服务相关功能
- 修改集合名称为 collection_8
- 将 ArchiveDto 中的 arcsoftFeature 字段类型从 byte[] 改为 float[]
- 更新 MilvusService 接口和实现类中的方法签名
- 修改向量搜索和索引创建的相关逻辑
-优化 MilvusUtil 中的 arcsoftToFloat 方法

Took 1 hour 1 minute

+++ src/main/java/org/alex/entity/ArchiveDto.java
@@ -14,7 +14,7 @@ public class ArchiveDto {
     private Integer orgId;
 
     private String archiveId;
-    private byte[] arcsoftFeature;
+    private float[] arcsoftFeature;
 
     private String text;
     private String fileName;
